Cleaning (remove audio and file as not supported yet)
